requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
SpeechRecognition==3.10.1
pydub==0.25.1
python-Levenshtein==0.23.0
numpy==1.26.2
librosa==0.10.1
